Extract Result from policy.go into its own file
